Rearranging files, fixing return codes
